requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Bootstrap==3.3.7.1
pysnmp==4.4.12
python-nmap==0.6.1
python-dotenv==0.19.2
SQLAlchemy==1.4.47
